Rename JiraSlashcommand to a proper name
