+++ slashcommands/jira.ts
@@ -3,11 +3,11 @@ import { IRoom } from '@rocket.chat/apps-engine/definition/rooms';
 import { ISlashCommand, SlashCommandContext } from '@rocket.chat/apps-engine/definition/slashcommands';
 import { URL } from 'url';
 
-import { CommandEnum } from './enums/CommandEnum';
-import { JiraApp } from './index';
-import { formatIssueMessage, startNewMessageWithDefaultSenderConfig } from './lib/helpers';
-import { getConnectedProjects, isProjectConnected, persistConnectedProjects } from './lib/persistence';
-import { IJiraError, IJiraIssue, sdk } from './lib/sdk';
+import { CommandEnum } from '../enums/CommandEnum';
+import { JiraApp } from '../index';
+import { formatIssueMessage, startNewMessageWithDefaultSenderConfig } from '../lib/helpers';
+import { getConnectedProjects, isProjectConnected, persistConnectedProjects } from '../lib/persistence';
+import { IJiraError, IJiraIssue, sdk } from '../lib/sdk';
 
 export class JiraSlashcommand implements ISlashCommand {
     public command = 'jira';
